+++ mipac/actions/admins/__init__.py
@@ -1,2 +1,4 @@
+from .admin import *
+from .announcement import *
 from .emoji import *
 from .moderator import *