+++ mipac/actions/mute.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, AsyncIterable
+from typing import TYPE_CHECKING, AsyncGenerator
 
 from mipac.abstract.action import AbstractAction
 from mipac.errors.base import ParameterError
@@ -14,13 +14,7 @@ if TYPE_CHECKING:
 
 
 class MuteActions(AbstractAction):
-    def __init__(
-        self,
-        user_id: str | None = None,
-        *,
-        session: HTTPClient,
-        client: ClientManager
-    ):
+    def __init__(self, user_id: str | None = None, *, session: HTTPClient, client: ClientManager):
         self._user_id: str | None = user_id
         self._session: HTTPClient = session
         self._client: ClientManager = client
@@ -75,22 +69,17 @@ class MuteActions(AbstractAction):
         since_id: str | None = None,
         until_id: str | None = None,
         all: bool = True,
-    ) -> AsyncIterable[MuteUser]:
+    ) -> AsyncGenerator[MuteUser, None]:
         if limit > 100:
             raise ParameterError('limit は100以下である必要があります')
 
         async def request(body) -> list[MuteUser]:
             res: list[IMuteUser] = await self._session.request(
-                Route('POST', '/api/mute/list'),
-                lower=True,
-                auth=True,
-                json=body,
+                Route('POST', '/api/mute/list'), lower=True, auth=True, json=body,
             )
             return [MuteUser(user, client=self._client) for user in res]
 
-        data = remove_dict_empty(
-            {'limit': limit, 'sinceId': since_id, 'untilId': until_id}
-        )
+        data = remove_dict_empty({'limit': limit, 'sinceId': since_id, 'untilId': until_id})
 
         if all:
             data['limit'] = 100