+++ mipac/actions/user.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, AsyncIterator, Literal, Optional
+from typing import TYPE_CHECKING, AsyncGenerator, Literal, Optional
 
 from mipac.errors.base import NotExistRequiredData, ParameterError
 from mipac.http import HTTPClient, Route
@@ -16,10 +16,7 @@ __all__ = ['UserActions']
 
 class UserActions:
     def __init__(
-        self,
-        session: HTTPClient,
-        client: ClientManager,
-        user: Optional[LiteUser] = None,
+        self, session: HTTPClient, client: ClientManager, user: Optional[LiteUser] = None,
     ):
         self.__session: HTTPClient = session
         self.__user: Optional[LiteUser] = user
@@ -36,9 +33,7 @@ class UserActions:
         return UserDetailed(res, client=self.__client)  # TODO: 自分用のクラスに変更する
 
     def get_profile_link(
-        self,
-        external: bool = True,
-        protocol: Literal['http', 'https'] = 'https',
+        self, external: bool = True, protocol: Literal['http', 'https'] = 'https',
     ):
         if not self.__user:
             return None
@@ -48,18 +43,13 @@ class UserActions:
             else self.__session._url
         )
         path = (
-            f'/@{self.__user.username}'
-            if external
-            else f'/{self.__user.api.action.get_mention()}'
+            f'/@{self.__user.username}' if external else f'/{self.__user.api.action.get_mention()}'
         )
         return host + path
 
     @cache(group='get_user')
     async def get(
-        self,
-        user_id: str | None = None,
-        username: str | None = None,
-        host: str | None = None,
+        self, user_id: str | None = None, username: str | None = None, host: str | None = None,
     ) -> UserDetailed:
         """
         ユーザーのプロフィールを取得します。一度のみサーバーにアクセスしキャッシュをその後は使います。
@@ -80,9 +70,7 @@ class UserActions:
             ユーザー情報
         """
 
-        field = remove_dict_empty(
-            {'userId': user_id, 'username': username, 'host': host}
-        )
+        field = remove_dict_empty({'userId': user_id, 'username': username, 'host': host})
         data = await self.__session.request(
             Route('POST', '/api/users/show'), json=field, auth=True, lower=True
         )
@@ -90,10 +78,7 @@ class UserActions:
 
     @cache
     async def fetch(
-        self,
-        user_id: str | None = None,
-        username: str | None = None,
-        host: str | None = None,
+        self, user_id: str | None = None, username: str | None = None, host: str | None = None,
     ) -> UserDetailed:
         """
         サーバーにアクセスし、ユーザーのプロフィールを取得します。基本的には get_userをお使いください。
@@ -115,9 +100,7 @@ class UserActions:
         if not check_multi_arg(user_id, username):
             raise ParameterError('user_id, usernameどちらかは必須です')
 
-        field = remove_dict_empty(
-            {'userId': user_id, 'username': username, 'host': host}
-        )
+        field = remove_dict_empty({'userId': user_id, 'username': username, 'host': host})
         data = await self.__session.request(
             Route('POST', '/api/users/show'), json=field, auth=True, lower=True
         )
@@ -179,11 +162,7 @@ class UserActions:
 
         if user is None:
             raise NotExistRequiredData('Required parameters: user')
-        return (
-            f'@{user.username}@{user.host}'
-            if user.instance
-            else f'@{user.username}'
-        )
+        return f'@{user.username}@{user.host}' if user.instance else f'@{user.username}'
 
     async def search(
         self,
@@ -194,7 +173,7 @@ class UserActions:
         detail: bool = True,
         *,
         all: bool = False,
-    ) -> AsyncIterator[UserDetailed | LiteUser]:
+    ) -> AsyncGenerator[UserDetailed | LiteUser, None]:
         """
         Search users by keyword.
 
@@ -215,8 +194,8 @@ class UserActions:
 
         Returns
         -------
-        AsyncIterator[UserDetailed | LiteUser]
-            A AsyncIterator of users.
+        AsyncGenerator[UserDetailed | LiteUser, None]
+            A AsyncGenerator of users.
         """
 
         if limit > 100:
@@ -224,10 +203,7 @@ class UserActions:
 
         async def request(body) -> list[UserDetailed | LiteUser]:
             res = await self.__session.request(
-                Route('POST', '/api/users/search'),
-                lower=True,
-                auth=True,
-                json=body,
+                Route('POST', '/api/users/search'), lower=True, auth=True, json=body,
             )
             return [
                 UserDetailed(user, client=self.__client)
@@ -237,13 +213,7 @@ class UserActions:
             ]
 
         body = remove_dict_empty(
-            {
-                'query': query,
-                'limit': limit,
-                'offset': offset,
-                'origin': origin,
-                'detail': detail,
-            }
+            {'query': query, 'limit': limit, 'offset': offset, 'origin': origin, 'detail': detail}
         )
 
         if all:
@@ -292,12 +262,7 @@ class UserActions:
             raise ParameterError('limit は100以下である必要があります')
 
         body = remove_dict_empty(
-            {
-                'username': username,
-                'host': host,
-                'limit': limit,
-                'detail': detail,
-            }
+            {'username': username, 'host': host, 'limit': limit, 'detail': detail}
         )
         res = await self.__session.request(
             Route('POST', '/api/users/search-by-username-and-host'),