+++ mipac/manager/admins/__init__.py
@@ -1,4 +1,5 @@
-from .emoji import *
 from .admin import *
+from .announcement import *
+from .emoji import *
 from .moderator import *
 from .user import *