+++ mipac/manager/client.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from mipac.actions.client import ClientActions
 
+from mipac.actions.client import ClientActions
 from mipac.http import HTTPClient
 from mipac.manager.admins.admin import AdminManager
 from mipac.manager.chart import ChartManager