+++ mipac/models/follow.py
@@ -11,9 +11,7 @@ if TYPE_CHECKING:
 
 
 class FollowRequest:
-    def __init__(
-        self, follow_request: IFollowRequest, *, client: ClientManager
-    ) -> None:
+    def __init__(self, follow_request: IFollowRequest, *, client: ClientManager) -> None:
         self.__follow_request = follow_request
         self.__client = client
 
@@ -23,18 +21,12 @@ class FollowRequest:
 
     @property
     def follower(self) -> LiteUser:
-        return LiteUser(
-            self.__follow_request['follower'], client=self.__client
-        )
+        return LiteUser(self.__follow_request['follower'], client=self.__client)
 
     @property
     def followee(self) -> LiteUser:
-        return LiteUser(
-            self.__follow_request['followee'], client=self.__client
-        )
+        return LiteUser(self.__follow_request['followee'], client=self.__client)
 
     @property
     def api(self) -> FollowRequestManager:
-        return self.__client._create_user_instance(
-            user=self.follower
-        ).follow.request
+        return self.__client._create_user_instance(user=self.follower).follow.request