+++ mipac/models/lite/emoji.py
@@ -1,17 +1,16 @@
 from __future__ import annotations
+
 from typing import TYPE_CHECKING
 
-from mipac.types.emoji import ICustomEmojiLite
 from mipac.config import config
+from mipac.types.emoji import ICustomEmojiLite
 
 if TYPE_CHECKING:
     from mipac.manager.client import ClientManager
 
 
 class PartialCustomEmoji:
-    def __init__(
-        self, emoji: ICustomEmojiLite, *, client: ClientManager
-    ) -> None:
+    def __init__(self, emoji: ICustomEmojiLite, *, client: ClientManager) -> None:
         self.__emoji = emoji
         self.__client: ClientManager = client