+++ mipac/types/__init__.py
@@ -1,4 +1,5 @@
 from .ads import *
+from .announcement import *
 from .channel import *
 from .chart import *
 from .chat import *
@@ -14,3 +15,4 @@ from .page import *
 from .poll import *
 from .reaction import *
 from .user import *
+from .admin import *